mysite/requirement.txt
asgiref==3.2.3
Django==3.0
django-AutoDateTimeFields==0.9.0
django-shortuuidfield==0.1.3
Pillow==6.2.1
psycopg2-binary==2.8.4
pytz==2019.3
shortuuid==0.5.0
six==1.13.0
sqlparse==0.3.0
